앨범 삭제 api 연동

+++ src/MenuComponents/Board/Album/AlbumDetail.js
@@ -9,6 +9,8 @@ const AlbumDetail = ({ onDelete, onEdit }) => {
    const navigate = useNavigate();
    const { albumId } = useParams(); 
    const [ album, setAlbum] = useState(null);
+   const [showConfirmation, setShowConfirmation] = useState(false);
+
 
    useEffect(() => {
       const fetchNoticeById = async () => {
@@ -37,12 +39,32 @@ const AlbumDetail = ({ onDelete, onEdit }) => {
       return <div>로딩 중...</div>;
     }
 
-
     const handleDelete = () => {
-      onDelete(albumId);
-      navigate('/Board/albums');
+      setShowConfirmation(true);
     };
   
+    const handleConfirmDelete = async () => {
+      try {
+        const deleteResponse = await axios.delete(`https://umust302.shop/api/articles/${albumId}`);
+        if (deleteResponse.status === 200) {
+          console.log('press deleted successfully.');
+          onDelete(albumId);
+  
+          navigate('/Board/albums');
+        } else {
+          console.error('Failed to delete album.');
+        }
+      } catch (error) {
+        console.error('Error deleting album:', error);
+      }
+  
+      setShowConfirmation(false);
+    };
+  
+  
+    const handleCancelDelete = () => {
+      setShowConfirmation(false);
+    };
     const handleEdit = () => {
       onEdit(albumId);
       navigate(`/Board/albums/${albumId}/edit`);
@@ -64,14 +86,26 @@ const AlbumDetail = ({ onDelete, onEdit }) => {
         )}
 
         <S.AlbumDetails>
-            <span>작성자: {album.createdBy || '알 수 없음'}</span>
+            <span>작성자: {album.createdBy || '짱구'}</span>
             <span>작성 시간: {(new Date(album.createdAt)).toLocaleString() || '알 수 없음'}</span>
         </S.AlbumDetails>
 
         <S.Buttons>
+        <button onClick={handleDelete}>삭제</button>
+        <button onClick={handleEdit}>수정</button>
+      </S.Buttons>
 
-        </S.Buttons>
-
+        {showConfirmation && (
+          <S.ConfirmationPopup>
+            <S.ConfirmationPopupContent>
+              <p>정말로 삭제하시겠습니까?</p>
+              <S.ConfirmationButtons>
+                <button onClick={handleConfirmDelete}>예</button>
+                <button onClick={handleCancelDelete}>아니요</button>
+              </S.ConfirmationButtons>
+            </S.ConfirmationPopupContent>
+          </S.ConfirmationPopup>
+        )}
          </S.DetailContainer>
       </S.AlbumDetailContainer>
    );
